examples/minimal-repo/app_server/pyproject.toml
[project]
name = "minimal-repo-app-server"
version = "0.1.0"
description = "App server with backup, analysis jobs, CSV export"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "httpx>=0.27.0",
  "sqlalchemy>=2.0.0",
  "aiosqlite>=0.20.0",
  "pandas>=2.2.0",
  "apscheduler>=3.10.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2.7.0",
  "openpyxl>=3.1.0",
  "tenacity>=9.0.0",
  "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.24.0",
]
